remove router
